Update tpch to 0.6 and change schemas to match
Schema changes for TPC-H tables that affect tests:
- Change linenumber from BIGINT to INTEGER
- Change shippriority from BIGINT to INTEGER
- Change quantity from BIGINT to DOUBLE
- Change partsupp.type from BIGINT to INTEGER

+++ presto-tests/src/main/java/com/facebook/presto/tests/AbstractTestIntegrationSmokeTest.java
@@ -313,7 +313,7 @@ public abstract class AbstractTestIntegrationSmokeTest
                     .row("orderdate", orderDateType, "")
                     .row("orderpriority", "varchar", "")
                     .row("clerk", "varchar", "")
-                    .row("shippriority", "bigint", "")
+                    .row("shippriority", "integer", "")
                     .row("comment", "varchar", "")
                     .build();
     }
